[ADD] get list of games, list of players in game and remove player in game when disconnected

+++ include/Network/Message.hpp
@@ -45,7 +45,7 @@ class Message {
             return _header.bodySize;
         }
 
-        void setResponseCode(int code)
+        void setResponseCode(short code)
         {
             _header.responseCode = code;
         }
